ivshmem-net: Fix bogus transition to RESET state
If we are in READY but the remote is still in INIT, we so far fell back
to RESET which caused the setup to get stuck. Fix this by only
transitioning from READY/RUN to RESET in ivshm_net_state_change if the
remote is in RESET as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
(cherry picked from commit ef53137ba67d41cade192299fad416165e3b07e9)
Signed-off-by: Khiem Nguyen <khiem.nguyen.xt@renesas.com>