ivshmem-net: Fix bogus transition to RESET state
If we are in READY but the remote is still in INIT, we so far fell back
to RESET which caused the setup to get stuck. Fix this by only
transitioning from READY/RUN to RESET in ivshm_net_state_change if the
remote is in RESET as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
(cherry picked from commit bbbd9b3f188c0ee9c93e5a450a81666d5117a216)
Signed-off-by: Khiem Nguyen <khiem.nguyen.xt@renesas.com>